Simple field update correctly typed.
Next challenge is to solve the "uninitialized fields" problem.
We start by typing main, therefore initializers are not typed yet: any
field that the initializer creates is therefore blank / null, and any
subsequent usage of said field brings a null exception.
Simplest solution is to provide temporary, "standard" fields at object
creation, even if it is not the most elegant.